pyproject.toml
[tool.poetry]
name = "itba-pda"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.9.7"
apache-airflow = "2.2.5"
python-decouple = "^3.6"
requests = "^2.28.1"
setuptools = "^65.3.0"
psycopg2 = "^2.9.3"
numpy = "^1.23.3"
pandas = "^1.5.0"
streamlit = "^1.13.0"
matplotlib = "^3.6.0"
Jinja2 = "3.0"
plotly-express = "^0.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.8.0"
pylint = "^2.15.0"
pre-commit = "^2.20.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
